+++ test/test_server_info.py
@@ -1,27 +1,34 @@
 import json
 import os
 
+from qgis.core import Qgis
+
 __copyright__ = 'Copyright 2021, 3Liz'
 __license__ = 'GPL version 3'
 __email__ = 'info@3liz.org'
 
+QUERY = "/lizmap/server.json"
+KEY = 'QGIS_SERVER_LIZMAP_REVEAL_SETTINGS'
+
 
 def test_lizmap_server_info(client):
     """Test the Lizmap API for server settings"""
-    query = "/lizmap/server.json"
-    key = 'QGIS_SERVER_LIZMAP_REVEAL_SETTINGS'
-
     # The environment variable is already there
-    assert os.getenv(key) == 'TRUE'
+    assert os.getenv(KEY) == 'TRUE'
 
     # The query must work
-    rv = client.get(query)
+    rv = client.get(QUERY)
     assert rv.status_code == 200
 
     assert rv.headers.get('Content-Type', '').find('application/json') == 0
 
     json_content = json.loads(rv.content.decode('utf-8'))
     assert 'qgis_server' in json_content
+
+    if 33000 <= Qgis.QGIS_VERSION_INT < 33200:
+        assert json_content['qgis_server']['metadata']['name'] == "'s-Hertogenbosch"
+    elif 32800 <= Qgis.QGIS_VERSION_INT < 32800:
+        assert json_content['qgis_server']['metadata']['name'] == "Firenze"
 
     # Names and versions are used in Lizmap Web Client
     expected_plugins = ('atlasprint', 'wfsOutputExtension', 'lizmap_server')
@@ -32,9 +39,12 @@ def test_lizmap_server_info(client):
 
     assert len(json_content['fonts']) >= 1
 
+
+def test_lizmap_server_info_env_check(client):
+    """ Check the environment variable check. """
     # Remove the security environment variable, the query mustn't work
-    del os.environ[key]
-    rv = client.get(query)
+    del os.environ[KEY]
+    rv = client.get(QUERY)
     assert rv.status_code == 400
 
     assert rv.headers.get('Content-Type', '').find('application/json') == 0
@@ -42,5 +52,5 @@ def test_lizmap_server_info(client):
     json_content = json.loads(rv.content.decode('utf-8'))
     assert [{'code': 'Bad request error', 'description': 'Invalid request'}] == json_content
 
-    # Reset the environment variable
-    os.environ[key] = 'TRUE'
+    # Reset the environment variable just in case
+    os.environ[KEY] = 'TRUE'